use Long ids

+++ hospital-jdbc/src/main/java/ir/maktab/hospital/dao/PatientDao.java
@@ -40,7 +40,7 @@ public class PatientDao implements BaseDao<PatientEntity> {
             ResultSet resultSet = ps.executeQuery();
 
             if (resultSet.next()) {
-                int id = resultSet.getInt(1);
+                Long id = resultSet.getLong(1);
                 String address = resultSet.getString(2);
                 String name = resultSet.getString(3);
                 int age = resultSet.getInt(4);
